Adding a new step for the method used for the dataset step
