+++ openlab/frontend/src/pages/indicator-specification-cards/creator/indicator-specification-card.jsx
@@ -5,9 +5,9 @@ import SpecifyRequirements from "./components/specify-requirements/specify-requi
 import ChoosePath from "./components/choose-path/choose-path.jsx";
 import Visualization from "./components/visualization/visualization.jsx";
 import Dataset from "./components/dataset/dataset.jsx";
-import Dataset2 from "./components/dataset/dataset2.jsx";
 import Finalize from "./components/finalize/finalize.jsx";
 import { useSnackbar } from "notistack";
+import Method from "./components/method/method.jsx";
 
 export const ISCContext = createContext(undefined);
 
@@ -126,9 +126,14 @@ const IndicatorSpecificationCard = () => {
           finalize: {
             locked: true,
             openPanel: false,
-            step: "5",
+            step: "6",
           },
-          method:{locked: true,type:""},
+          method:{
+            locked: true,
+            type:"",
+            openPanel: false,
+            step:"0"
+          },
           manual:{
             locked:true,
             openPanel: false,
@@ -219,15 +224,27 @@ const IndicatorSpecificationCard = () => {
           )}
           {lockedStep.dataset.step === "4" && (
             <Grid size={{ xs: 12 }}>
-              <Dataset2/>
+              <Dataset/>
             </Grid>
           )}
+          {/* Adding the new step "Method" */}
+          {lockedStep.dataset.step === '4' && lockedStep.method.type !== ""  &&(
+            <Grid size={{ xs: 12 }}>
+              <Method />
+            </Grid>
+          )}
+
           {lockedStep.dataset.step === "3" && (
             <Grid size={{ xs: 12 }}>
-              <Dataset2 />
+              <Dataset />
             </Grid>
           )}
-          {lockedStep.visualization.step === "4" && (
+          {lockedStep.dataset.step === '3' && lockedStep.method.type !== ""  &&(
+            <Grid size={{ xs: 12 }}>
+              <Method />
+            </Grid>
+          )}
+          {lockedStep.visualization.step === "5" && (
             <Grid size={{ xs: 12 }}>
               <Visualization />
             </Grid>
